Added Deadline mode
Copied the Evolved mode and added a timer (180 seconds). Also changed
the updateParallax on the game mode to just be an update function as it
can be used to check lives and such later.

It's currently hard coded to launch Deadline. I will add the menu system
next.

Removed "shapeshifter.ogg", going to stick with the deadline music so
now there's "deadline.ogg"