Added new menu
Added a new select menu as well as updated other menu systems so a
player can go straight to the select screen. I was going to add left and
right navigation on the select screen by simply adding 2 to the
selectedItem, hoever, it would be possible to move 2 items on every menu
then.

Also added 3 new graphics for the menu.